Test role with playground (#2)
Use now the https://github.com/Tecnativa/ansible-test-playground
framework to test, which lets tests use systemd, and will also be
needed in the future to test swarm deployment.

Of course, it includes fixes for the detected failures, and it removes
the testing mode, since now we can run a full test.

Some noticeable changes:

- To move files, we copy and remove instead of hardlinking and
  removing, because hardlinking can fail when found on different
  partitions, which is quite common in `/tmp`.
- Workaround for ansible/ansible#31344.
- Support and test python 2 in hosts.
- Use `vfs` storage driver in tests; `overlay2` failed.
- Allow customizable bootloader.
- Disable bootloader stuff in tests.